Add ECR client.
I've done this client in slightly different way. It uses the AWS async
non blocking client to get the results, uses Monix to convert this to a
cats-effect IO monad. This means the calling code can be pure and
functional but we can still do it in a non blocking async way.