Add kms client and decryption utils.
We need to be able to decrypt things encrypted by kms as we're going to
start encrypting lambda variables in transit.

There are two methods here. One which just returns a single encrypted
value and one that takes a list of config paths and returns a map of
configPath -> decryptedValue. This is useful for some of the lambdas
which quite a few environment variables and reduces the code changes
needed in the other projects.