Add method which resets SQS message visibility
This is useful when an SQS message processor encounters an error,
so the message needs to be retried.

If the message visibility is not reset, the message will only become
visible when its visibility timeout expires. If the processor calls
this new method, the message will be available for retry immediately.