Add send heartbeat task function.
I've added in some tests for the function. I've also upgraded the scala version as I think we were running into this bug.
https://github.com/scala/bug/issues/12038
There was a similar error on one of the AWS classes which had a builder and only happened every second test run. It's gone away with the upgrade.