Remove the monix library.
I want to start migrating the cats-effect projects to cats-effect 3.
This one is using monix which hasn't yet moved over to version 3 and so
I can't use version 3 in any projects which needs this library.

The easiest solution is to take it out as this can be done with the
scala standard converters and vanilla cats-effect.