Update test to use dummy credentials.
Because the test is calling the real presign method, it needed
credentials in the environment to work. This wasn't a problem on Jenkins
because they were there but now they won't be on GitHub actions.

Iv'e added in an apply method to pass in a presigner if we want to and
then set dummy credentials in the tests.

This should now run without AWS env vars being set.