Move to individual modules.
This library is very large when included in other projects because we're
pulling in so many of the AWS SDKs and mostly we don't need them all.

This splits out the libraries into individual modules. The classes have
mostly been moved around and each module has its own client class.

They are all released together. I've done it this way to save time and
because most of the updates we get for this project are dependency
updates to cats which affects all of the modules anyway.